feat(border): support decimal border widths and radii

+++ modules/ensemble/lib/framework/model.dart
@@ -97,37 +97,37 @@ class BackgroundImage {
 
 class EBorderRadius {
   EBorderRadius._(
-      int _topLeft, int _topRight, int _bottomRight, int _bottomLeft)
+      double _topLeft, double _topRight, double _bottomRight, double _bottomLeft)
       : topLeft =
-            _topLeft == 0 ? Radius.zero : Radius.circular(_topLeft.toDouble()),
+            _topLeft == 0 ? Radius.zero : Radius.circular(_topLeft),
         topRight = _topRight == 0
             ? Radius.zero
-            : Radius.circular(_topRight.toDouble()),
+            : Radius.circular(_topRight),
         bottomRight = _bottomRight == 0
             ? Radius.zero
-            : Radius.circular(_bottomRight.toDouble()),
+            : Radius.circular(_bottomRight),
         bottomLeft = _bottomLeft == 0
             ? Radius.zero
-            : Radius.circular(_bottomLeft.toDouble());
+            : Radius.circular(_bottomLeft);
 
   Radius topLeft, topRight, bottomRight, bottomLeft;
 
-  factory EBorderRadius.all(int val) {
+  factory EBorderRadius.all(double val) {
     return EBorderRadius._(val, val, val, val);
   }
   // first value: top-left & bottom-right
   // second value: top-right & bottom-left
-  factory EBorderRadius.two(int first, int second) {
+  factory EBorderRadius.two(double first, double second) {
     return EBorderRadius._(first, second, first, second);
   }
   // first value: top-left
   // second value: top-right & bottom-left
   // third value: bottom-right
-  factory EBorderRadius.three(int first, int second, int third) {
+  factory EBorderRadius.three(double first, double second, double third) {
     return EBorderRadius._(first, second, third, second);
   }
   factory EBorderRadius.only(
-      int topLeft, int topRight, int bottomRight, int bottomLeft) {
+      double topLeft, double topRight, double bottomRight, double bottomLeft) {
     return EBorderRadius._(topLeft, topRight, bottomRight, bottomLeft);
   }
